% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_sales_table.R
\name{mod_sales_table_ui}
\alias{mod_sales_table_ui}
\title{Sales Table Module UI}
\usage{
mod_sales_table_ui(id)
}
\arguments{
\item{id}{Module namespace ID.}
}
\value{
A `tableOutput` element.
}
\description{
Places the `tableOutput` element for the sales data table.
}
