% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_value_boxes.R
\name{mod_value_boxes_ui}
\alias{mod_value_boxes_ui}
\title{Value Boxes Module UI}
\usage{
mod_value_boxes_ui(id, interval_ms = 5000)
}
\arguments{
\item{id}{Module namespace ID.}

\item{interval_ms}{Poll interval in milliseconds, used to label the
"Checks Performed" box. Default `5000`.}
}
\value{
A `layout_columns` element containing four value boxes.
}
\description{
Renders four [bslib::value_box()] elements showing poll statistics:
checks performed, fetches performed, time of last check, and time of last
fetch. The Checks box also displays the configured poll cadence.
}
